backend/requirements.txt
asgiref==3.2.7
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.15.0
chardet==3.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==2.2.13
django-allauth==0.50.0
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-cors-headers==3.7.0
django-extensions==2.2.9
django-rest-knox==4.1.0
django-webpack-loader==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
Faker==4.1.0
filelock==3.4.2
h11==0.13.0
idna==2.9
importlib-metadata==1.6.1
mysql-connector-python==8.0.29
mysqlclient==2.0.3
oauthlib==3.2.0
outcome==1.1.0
Pillow==9.1.0
platformdirs==2.4.1
pycparser==2.20
PyJWT==2.4.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.1
selenium==4.1.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.25.9
virtualenv==20.13.0
wsproto==1.1.0
zipp==3.1.0
